More character controller work
Still missing object destruction upgrades, upgrade tokens/station and climbing.
Known issue - If you enter a hot or cold zone without the appropriate upgrade, the position the game places you in is still inside the hot/cold zone, so you're unable to move. Just need to add a timer or move the player a bit further away to fix this, will do tomorrow but tired now.